Изменила добавление, изменение, удаление привычки, использовала кнопки инлайн...

See merge request natalia_pikaleva/telegram-bot-healthy-habits!7

+++ bot/handlers/default_handlers/delete_habit.py
@@ -8,26 +8,36 @@ import requests
 from config import setup_logging, API_HOST
 import logging
 from collections import defaultdict
-from .get_habits import user_selected_habit
+from .get_habit import user_selected_habit
 
 setup_logging()
 logger = logging.getLogger(__name__)
 
 
-@bot.message_handler(func=lambda message: message.text == "Удалить привычку")
-def bot_delete_habit(message: Message) -> None:
-    """
-    Хендлер для удаления привычки
-    """
-    logger.info("Start bot_delete_habit")
+@bot.callback_query_handler(func=lambda call: call.data.startswith('delete_'))
+def callback_update_repeat_period(call):
+    habit_id = int(call.data.split('_')[1])
+    user_id = call.from_user.id
+    user_selected_habit[user_id]["habit_id"] = habit_id
+    if user_id not in user_selected_habit:
+        bot.send_message(user_id, "Выберите привычку из списка", reply_markup=habits_commands())
 
+    msg = bot.send_message(user_id, "Вы уверены, что хотите удалить привычку? Что удалить введите 1")
+    bot.register_next_step_handler(msg, process_habit_delete)
+
+
+def process_habit_delete(message):
     with SessionLocal() as db:
         token = get_token_for_user(db, message.chat.id)
     if not token:
         bot.send_message(message.chat.id, "Пожалуйста, авторизуйтесь через /start")
         return
 
+    if message.text != "1":
+        return
+
     habit_id = user_selected_habit[message.chat.id]["habit_id"]
+    bot.send_message(message.chat.id, f"Сейчас сохранена привычка с id: {user_selected_habit[message.chat.id]["habit_id"]}")
 
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.delete(f"http://{API_HOST}:8000/habits/{habit_id}/delete", headers=headers)
@@ -48,6 +58,6 @@ def bot_delete_habit(message: Message) -> None:
     if response.status_code == 200:
         data = response.json()
         # Обработка успешного ответа
-        bot.send_message(message.chat.id, f"data: {data}")
+        bot.send_message(message.chat.id, "Привычка удалена", reply_markup=habits_commands())
     else:
         bot.send_message(message.chat.id, "Ошибка при выполнении запроса.")
